move serve_static and shell_welcome_message into misc
