+++ archivebox/core/urls.py
@@ -5,9 +5,10 @@ from django.views import static
 from django.conf import settings
 from django.views.generic.base import RedirectView
 
-from .admin import archivebox_admin
-from .views import HomepageView, SnapshotView, PublicIndexView, AddView, HealthCheckView
-from .serve_static import serve_static
+from archivebox.misc.serve_static import serve_static
+
+from core.admin_site import archivebox_admin
+from core.views import HomepageView, SnapshotView, PublicIndexView, AddView, HealthCheckView
 
 # GLOBAL_CONTEXT doesn't work as-is, disabled for now: https://github.com/ArchiveBox/ArchiveBox/discussions/1306
 # from archivebox.config import VERSION, VERSIONS_AVAILABLE, CAN_UPGRADE
